API reference

Index

FastPriorityQueues.HeapPriorityQueue

FastPriorityQueues.SortedVectorPriorityQueue

FastPriorityQueues.VectorPriorityQueue

DataStructures.dequeue!

DataStructures.dequeue!

DataStructures.dequeue!

DataStructures.dequeue_pair!

DataStructures.dequeue_pair!

DataStructures.dequeue_pair!

DataStructures.enqueue!

DataStructures.enqueue!

DataStructures.enqueue!

Docstrings

FastPriorityQueues.HeapPriorityQueue — Type

HeapPriorityQueue{K,V}

Min priority queue with keys of type K and priority values of type V, stored using a binary heap from DataStructures.jl.

Fields

heap::BinaryHeap{K,V}: heap of key-value pairs k => v ordered by increasing v

FastPriorityQueues.SortedVectorPriorityQueue — Type

SortedVectorPriorityQueue{K,V}

Min priority queue with keys of type K and priority values of type V, stored using a sorted vector of couples.

Fields

pairs::Vector{Pair{K,V}}: vector of key-value pairs k => v ordered by increasing v.

FastPriorityQueues.VectorPriorityQueue — Type

VectorPriorityQueue{K,V}

Min priority queue with keys of type K and priority values of type V, stored as a vector of couples.

Fields

pairs::Vector{Pair{K,V}}: vector of key-value pairs k => v in arbitrary order.

DataStructures.dequeue! — Method

dequeue!(pq::HeapPriorityQueue)

Remove and return the key k with lowest priority value v. Amortized complexity O(1).

DataStructures.dequeue! — Method

dequeue!(pq::SortedVectorPriorityQueue)

Remove and return the key k with lowest priority value v. Amortized complexity O(1).

DataStructures.dequeue! — Method

dequeue!(pq::VectorPriorityQueue)

Remove and return the key k with lowest priority value v. Amortized complexity O(n).

DataStructures.dequeue_pair! — Method

dequeue_pair!(pq::HeapPriorityQueue)

Remove and return the pair k => v with lowest priority value v. Amortized complexity O(1).

DataStructures.dequeue_pair! — Method

dequeue_pair!(pq::SortedVectorPriorityQueue)

Remove and return the pair k => v with lowest priority value v. Amortized complexity O(1).

DataStructures.dequeue_pair! — Method

dequeue_pair!(pq::VectorPriorityQueue)

Remove and return the pair k => v with lowest priority value v. Amortized complexity O(n).

DataStructures.enqueue! — Method

enqueue!(pq::HeapPriorityQueue, k, v)

Insert k => v into the queue pq. Amortized complexity O(log n).

DataStructures.enqueue! — Method

enqueue!(pq::SortedVectorPriorityQueue, k, v)

Insert k => v into the queue pq. Amortized complexity O(log n).

DataStructures.enqueue! — Method

enqueue!(pq::VectorPriorityQueue, k, v)

Insert k => v into the queue pq. Amortized complexity O(1).